fix: req method checks, removed redundant testing, renamed root
