--- notes.md
@@ -1,5 +1,0 @@
-Serve relative paths
-Frontend can handle the domain and API urls.
-Path traversal vulns. 
-Do security checks.
-Try to create separate sessions for different users. UUID add.
